refactor(gdn): merge gdn_decode_bf16_state_wide_vec.py into the main BF16 file
Now that the BF16 surface is exactly two ``@cute.kernel``s
(``gdn_wide_vec_kernel`` and ``gdn_decode_bf16state_mtp_ilp4_kernel``),
keeping them in two separate Python files added more friction than
isolation. Both call into the same dispatcher, share the FMA helpers,
and need the same ``NUM_SMS`` / ``_USE_PACKED_FMA`` runtime constants.

Merges the wide_vec file into ``flashinfer/gdn_kernels/gdn_decode_bf16_state.py``:

- Wide_vec layout constants (``LANES_PER_ROW``, ``ELEMS_PER_LANE``,
  ``NUM_WARPS``, ``NUM_THREADS``, ``NUM_GROUPS``, ``ILP_ROWS``) inlined
  next to the MTP constants. Drops the unused ``TILE_K = 128``
  documentation constant from the wide_vec block.
- ``gdn_wide_vec_kernel`` and its ``_run_wide_vec`` launcher land in
  their own section in the merged file. Section ordering is now:
  wide_vec kernel + launcher, then mtp_ilp4 kernel + launcher.
- ``_compiled_kernels_wide_vec`` cache lives next to
  ``_compiled_kernels_mtp``.
- ``gated_delta_rule_mtp_wide_vec`` Python entry sits between the
  helper functions and the public dispatchers ``gated_delta_rule`` /
  ``gated_delta_rule_mtp``. Drops the two cross-file lazy imports
  (``from .gdn_decode_bf16_state_wide_vec import …``) that the
  dispatchers used to do — same-module references now.
- Drops the duplicated ``fma_pair_mul`` / ``fma_pair`` definitions and
  ``NUM_SMS`` / ``_GPU_MAJOR`` / ``_USE_PACKED_FMA`` constants that
  appeared in both files; the merged file uses the existing main-file
  definitions.

Test ``test_gdn_decode_bf16_state_wide_vec_mtp_kernel`` updated to
import ``gated_delta_rule_mtp_wide_vec`` from the merged location.

The two old files (1221 + 799 = 2020 lines) become one file at 1973
lines (~47 LOC saved from dedup of imports, FMA helpers, runtime
constants, and section headers).

No behavior change. Full BF16 / wide_vec test sweep verified prior to
merge; cross-shape correctness (B in {1, 8, 32} × T in {2, 4})
re-checked on the merged file (single == split bit-equivalent at
bf16 noise floor). Pre-commit clean.

AI-assisted by Claude Code.
